services/data-ingestion/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.1
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
textblob==0.17.1
APScheduler==3.10.4
jsonschema==4.20.0
python-json-logger==2.0.7
beautifulsoup4==4.12.2
